+++ CHANGELOG.md
@@ -15,7 +15,8 @@
 ## 1.76.2
 
 ### Features
-- limits: get block storage volume limit #577 
+- compute: Add Block Storage #574
+- limits: get block storage volume limit #577
 
 ### Improvements
 - x: make `exo api` an alias to `exo x` #579